+++ src/main/java/hackclient/rise/sk.java
@@ -29,6 +29,6 @@ public class sk implements Comparator<Integer> {
     }
 
     public int compare(Integer var1, Integer var2) {
-        return this.a((Integer)var1, (Integer)var2);
+        return this.a(var1, var2);
     }
 }
